+++ packages/spindle-hooks/src/useTimeDistance/index.ts
@@ -1,2 +1,1 @@
-export { useCarousel } from './useCarousel';
 export { useTimeDistance } from './useTimeDistance';